+++ frontends/PyCDE/test/behavioral.py
@@ -1,16 +1,13 @@
 # RUN: rm -rf %t
 # RUN: %PYTHON% %s %t 2>&1 | FileCheck %s
 
-from re import A
-from pycde import module, generator, types, Input, Output
-from pycde.constructs import If, Else, Then
-import pycde
-
-import sys
+from pycde import generator, types, Input, Output
+from pycde.behavioral import If, Else, EndIf
+from pycde.testing import unittestmodule
 
 
-@module
-class IfDemo:
+@unittestmodule()
+class IfNestedTest:
   cond = Input(types.i1)
   cond2 = Input(types.i1)
   a = Input(types.ui8)
@@ -23,22 +20,23 @@ class IfDemo:
   def build(ports):
     w = ports.a * ports.b
     with If(ports.cond):
-      with Then:
-        with If(ports.cond2):
-          with Then:
-            x = ports.a
-          with Else:
-            x = ports.b
-        v = ports.a * x
-        u = v * ports.b
+      with If(ports.cond2):
+        x = ports.a
       with Else:
-        v = ports.b.as_uint(16)
-        u = v * ports.a
+        x = ports.b
+      EndIf()
+      v = ports.a * x
+      u = v * ports.b
+    with Else:
+      v = ports.b.as_uint(16)
+      u = v * ports.a
+    EndIf()
+
     ports.out2 = u
     ports.out = v + w
 
 
-# CHECK-LABEL: msft.module @IfDemo {} (%a: ui8, %b: ui8, %cond: i1, %cond2: i1) -> (out: ui17, out2: ui24)
+# CHECK-LABEL: msft.module @IfNestedTest {} (%a: ui8, %b: ui8, %cond: i1, %cond2: i1) -> (out: ui17, out2: ui24)
 # CHECK:         %0 = hwarith.mul %a, %b {sv.namehint = "a_mul_b"} : (ui8, ui8) -> ui16
 # CHECK:         %1 = comb.mux %cond2, %a, %b {sv.namehint = "x"} : ui8
 # CHECK:         %2 = hwarith.mul %a, %1 {sv.namehint = "v_thenvalue"} : (ui8, ui8) -> ui16
@@ -50,7 +48,24 @@ class IfDemo:
 # CHECK:         %8 = hwarith.add %6, %0 {sv.namehint = "v_plus_a_mul_b"} : (ui16, ui16) -> ui17
 # CHECK:         msft.output %8, %7 : ui17, ui24
 
-t = pycde.System([IfDemo], name="BehavioralTest", output_directory=sys.argv[1])
-t.generate()
-t.print()
-# t.emit_outputs()
+
+@unittestmodule()
+class IfDefaultTest:
+  cond = Input(types.i1)
+  a = Input(types.ui8)
+  b = Input(types.ui8)
+
+  out = Output(types.ui8)
+
+  @generator
+  def build(ports):
+    v = ports.a
+    with If(ports.cond):
+      v = ports.b
+    EndIf()
+    ports.out = v
+
+
+# CHECK-LABEL: msft.module @IfDefaultTest {} (%a: ui8, %b: ui8, %cond: i1) -> (out: ui8)
+# CHECK:         [[r0:%.+]] = comb.mux %cond, %b, %a {sv.namehint = "v"} : ui8
+# CHECK:         msft.output [[r0]] : ui8
